Semi-Working Lexer
The Ragel code will now tokenise a buffer into integers. When the end
of the buffer is reached it repeatedly returns the end of file
token. Could look at tokenising the buffer into a quque of tokens and
then pickig one of these at a time rather than stopping the lexer each
time we reach a token.